backend/requirements.txt
Django~=5.0
python-dotenv~=1.0
psycopg2-binary
djangorestframework
networkx